+++ packages/openai/src/responses/openai-responses-api-types.ts
@@ -10,6 +10,7 @@ export type OpenAIResponsesMessage =
   | OpenAIResponsesFunctionCallOutput
   | OpenAIWebSearchCall
   | OpenAIComputerCall
+  | OpenAIFileSearchCall
   | OpenAIResponsesReasoning;
 
 export type OpenAIResponsesSystemMessage = {
@@ -32,6 +33,7 @@ export type OpenAIResponsesAssistantMessage = {
     | { type: 'output_text'; text: string }
     | OpenAIWebSearchCall
     | OpenAIComputerCall
+    | OpenAIFileSearchCall
   >;
   id?: string;
 };
@@ -62,6 +64,12 @@ export type OpenAIComputerCall = {
   status?: string;
 };
 
+export type OpenAIFileSearchCall = {
+  type: 'file_search_call';
+  id: string;
+  status?: string;
+};
+
 export type OpenAIResponsesTool =
   | {
       type: 'function';
@@ -84,7 +92,7 @@ export type OpenAIResponsesTool =
       vector_store_ids?: string[];
       max_num_results?: number;
       ranking_options?: {
-        ranker?: 'auto' | 'keyword' | 'semantic';
+        ranker?: 'auto' | 'default-2024-08-21';
       };
       filters?:
         | {