+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,260 @@
 # @ai-sdk/azure
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [63e2016]
+  - @ai-sdk/openai@2.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [bc45e29]
+  - @ai-sdk/openai@2.0.1
+
+## 2.0.0
+
+### Major Changes
+
+- d5f588f: AI SDK 5
+- 516be5b: ### Move Image Model Settings into generate options
+
+  Image Models no longer have settings. Instead, `maxImagesPerCall` can be passed directly to `generateImage()`. All other image settings can be passed to `providerOptions[provider]`.
+
+  Before
+
+  ```js
+  await generateImage({
+    model: luma.image('photon-flash-1', {
+      maxImagesPerCall: 5,
+      pollIntervalMillis: 500,
+    }),
+    prompt,
+    n: 10,
+  });
+  ```
+
+  After
+
+  ```js
+  await generateImage({
+    model: luma.image('photon-flash-1'),
+    prompt,
+    n: 10,
+    maxImagesPerCall: 5,
+    providerOptions: {
+      luma: { pollIntervalMillis: 5 },
+    },
+  });
+  ```
+
+  Pull Request: https://github.com/vercel/ai/pull/6180
+
+### Patch Changes
+
+- c467b38: feat (provider/azure): add OpenAI responses API support
+- 004fb17: feat(azure): add speech model support and fix transcription
+- 3b13ccf: feat(providers/azure): add transcribe
+- e2aceaf: feat: add raw chunk support
+- eb173f1: chore (providers): remove model shorthand deprecation warnings
+- db72adc: chore(providers/openai): update completion model to use providerOptions
+- 9bf7291: chore(providers/openai): enable structuredOutputs by default & switch to provider option
+- 9301f86: refactor (image-model): rename `ImageModelV1` to `ImageModelV2`
+- 3db5258: update: Azure OpenAI provider to v1 API format
+- 4617fab: chore(embedding-models): remove remaining settings
+- 52ce942: chore(providers/openai): remove & enable strict compatibility by default
+- 205077b: fix: improve Zod compatibility
+- Updated dependencies [a571d6e]
+- Updated dependencies [742b7be]
+- Updated dependencies [948b755]
+- Updated dependencies [d63bcbc]
+- Updated dependencies [e7fcc86]
+- Updated dependencies [7cddb72]
+- Updated dependencies [ccce59b]
+- Updated dependencies [e2b9e4b]
+- Updated dependencies [95857aa]
+- Updated dependencies [45c1ea2]
+- Updated dependencies [6f6bb89]
+- Updated dependencies [060370c]
+- Updated dependencies [dc714f3]
+- Updated dependencies [b5da06a]
+- Updated dependencies [d1a1aa1]
+- Updated dependencies [63f9e9b]
+- Updated dependencies [3bd3c0b]
+- Updated dependencies [5d142ab]
+- Updated dependencies [d5f588f]
+- Updated dependencies [e025824]
+- Updated dependencies [cc62234]
+- Updated dependencies [0571b98]
+- Updated dependencies [5d959e7]
+- Updated dependencies [b6b43c7]
+- Updated dependencies [4fef487]
+- Updated dependencies [48d257a]
+- Updated dependencies [0c0c0b3]
+- Updated dependencies [0d2c085]
+- Updated dependencies [0eee6a8]
+- Updated dependencies [177526b]
+- Updated dependencies [2f542fa]
+- Updated dependencies [40acf9b]
+- Updated dependencies [c15dfbf]
+- Updated dependencies [9222aeb]
+- Updated dependencies [3b1ea10]
+- Updated dependencies [e2aceaf]
+- Updated dependencies [d2af019]
+- Updated dependencies [411e483]
+- Updated dependencies [8ba77a7]
+- Updated dependencies [eb173f1]
+- Updated dependencies [209256d]
+- Updated dependencies [faea29f]
+- Updated dependencies [7032dc5]
+- Updated dependencies [870c5c0]
+- Updated dependencies [db72adc]
+- Updated dependencies [7b3ae3f]
+- Updated dependencies [a166433]
+- Updated dependencies [26735b5]
+- Updated dependencies [443d8ec]
+- Updated dependencies [a8c8bd5]
+- Updated dependencies [abf9a79]
+- Updated dependencies [14c9410]
+- Updated dependencies [e86be6f]
+- Updated dependencies [8d12da5]
+- Updated dependencies [9bf7291]
+- Updated dependencies [2e13791]
+- Updated dependencies [9f95b35]
+- Updated dependencies [d521cda]
+- Updated dependencies [66962ed]
+- Updated dependencies [0d06df6]
+- Updated dependencies [442be08]
+- Updated dependencies [472524a]
+- Updated dependencies [0059ee2]
+- Updated dependencies [dd3ff01]
+- Updated dependencies [8493141]
+- Updated dependencies [d9c98f4]
+- Updated dependencies [05d2819]
+- Updated dependencies [9301f86]
+- Updated dependencies [0a87932]
+- Updated dependencies [c4a2fec]
+- Updated dependencies [957b739]
+- Updated dependencies [79457bd]
+- Updated dependencies [a3f768e]
+- Updated dependencies [7435eb5]
+- Updated dependencies [8aa9e20]
+- Updated dependencies [4617fab]
+- Updated dependencies [b5a0e32]
+- Updated dependencies [136819b]
+- Updated dependencies [516be5b]
+- Updated dependencies [ac34802]
+- Updated dependencies [0054544]
+- Updated dependencies [cb68df0]
+- Updated dependencies [52ce942]
+- Updated dependencies [db64cbe]
+- Updated dependencies [ad80501]
+- Updated dependencies [b3c3450]
+- Updated dependencies [68ecf2f]
+- Updated dependencies [9e9c809]
+- Updated dependencies [32831c6]
+- Updated dependencies [48249c4]
+- Updated dependencies [c7d3b2e]
+- Updated dependencies [6dc848c]
+- Updated dependencies [6b98118]
+- Updated dependencies [d0f9495]
+- Updated dependencies [63d791d]
+- Updated dependencies [87b828f]
+- Updated dependencies [3f2f00c]
+- Updated dependencies [bfdca8d]
+- Updated dependencies [0ff02bb]
+- Updated dependencies [7979f7f]
+- Updated dependencies [ad2a3d5]
+- Updated dependencies [9943464]
+- Updated dependencies [39a4fab]
+- Updated dependencies [44f4aba]
+- Updated dependencies [0fa7414]
+- Updated dependencies [9bd5ab5]
+- Updated dependencies [57edfcb]
+- Updated dependencies [faf8446]
+- Updated dependencies [efc3a62]
+- Updated dependencies [7ea4132]
+- Updated dependencies [fa758ea]
+- Updated dependencies [d1a034f]
+- Updated dependencies [5c56081]
+- Updated dependencies [fd65bc6]
+- Updated dependencies [023ba40]
+- Updated dependencies [e497698]
+- Updated dependencies [ea7a7c9]
+- Updated dependencies [26535e0]
+- Updated dependencies [928fadf]
+- Updated dependencies [e030615]
+- Updated dependencies [5e57fae]
+- Updated dependencies [0a87932]
+- Updated dependencies [393138b]
+- Updated dependencies [c57e248]
+- Updated dependencies [5147e6e]
+- Updated dependencies [88a8ee5]
+- Updated dependencies [06bac05]
+- Updated dependencies [41fa418]
+- Updated dependencies [205077b]
+- Updated dependencies [71f938d]
+- Updated dependencies [3795467]
+- Updated dependencies [28a5ed5]
+- Updated dependencies [7182d14]
+- Updated dependencies [c1e6647]
+- Updated dependencies [c2b92cc]
+- Updated dependencies [284353f]
+- Updated dependencies [1766ede]
+- Updated dependencies [6f231db]
+- Updated dependencies [811dff3]
+- Updated dependencies [f10304b]
+- Updated dependencies [dd5fd43]
+- Updated dependencies [4af5233]
+- Updated dependencies [33f4a6a]
+- Updated dependencies [383cbfa]
+- Updated dependencies [27deb4d]
+- Updated dependencies [7df7a25]
+- Updated dependencies [c4df419]
+  - @ai-sdk/provider-utils@3.0.0
+  - @ai-sdk/provider@2.0.0
+  - @ai-sdk/openai@2.0.0
+
+## 2.0.0-beta.16
+
+### Patch Changes
+
+- Updated dependencies [88a8ee5]
+  - @ai-sdk/provider-utils@3.0.0-beta.10
+  - @ai-sdk/openai@2.0.0-beta.16
+
+## 2.0.0-beta.15
+
+### Patch Changes
+
+- Updated dependencies [9943464]
+- Updated dependencies [27deb4d]
+  - @ai-sdk/openai@2.0.0-beta.15
+  - @ai-sdk/provider@2.0.0-beta.2
+  - @ai-sdk/provider-utils@3.0.0-beta.9
+
+## 2.0.0-beta.14
+
+### Patch Changes
+
+- 004fb17: feat(azure): add speech model support and fix transcription
+- eb173f1: chore (providers): remove model shorthand deprecation warnings
+- Updated dependencies [eb173f1]
+- Updated dependencies [7032dc5]
+- Updated dependencies [dd5fd43]
+  - @ai-sdk/openai@2.0.0-beta.14
+  - @ai-sdk/provider-utils@3.0.0-beta.8
+
+## 2.0.0-beta.13
+
+### Patch Changes
+
+- Updated dependencies [e7fcc86]
+  - @ai-sdk/provider-utils@3.0.0-beta.7
+  - @ai-sdk/openai@2.0.0-beta.13
+
 ## 2.0.0-beta.12
 
 ### Patch Changes