+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,279 @@
 # @ai-sdk/google-vertex
 
+## 3.0.2
+
+### Patch Changes
+
+- b9cd900: feat(providers/google-vertex) Add TaskType support for Text Embedding Model
+- Updated dependencies [f5464aa]
+  - @ai-sdk/google@2.0.1
+
+## 3.0.1
+
+### Patch Changes
+
+- 11e3ba4: Make revisedPrompt nullish in schema
+
+## 3.0.0
+
+### Major Changes
+
+- d5f588f: AI SDK 5
+- 516be5b: ### Move Image Model Settings into generate options
+
+  Image Models no longer have settings. Instead, `maxImagesPerCall` can be passed directly to `generateImage()`. All other image settings can be passed to `providerOptions[provider]`.
+
+  Before
+
+  ```js
+  await generateImage({
+    model: luma.image('photon-flash-1', {
+      maxImagesPerCall: 5,
+      pollIntervalMillis: 500,
+    }),
+    prompt,
+    n: 10,
+  });
+  ```
+
+  After
+
+  ```js
+  await generateImage({
+    model: luma.image('photon-flash-1'),
+    prompt,
+    n: 10,
+    maxImagesPerCall: 5,
+    providerOptions: {
+      luma: { pollIntervalMillis: 5 },
+    },
+  });
+  ```
+
+  Pull Request: https://github.com/vercel/ai/pull/6180
+
+### Minor Changes
+
+- 6ca44f2: Fixed global region for vertex provider
+
+### Patch Changes
+
+- 8e171f5: feat (provider/google-vertex): add imagen-3.0-generate-002
+- cea5997: chore(providers/google-vertex): update embedding model to use providerOptions
+- 9ccce3a: feat (google-vertex): Set `.providerMetaData` for image model responses
+- e2aceaf: feat: add raw chunk support
+- 26735b5: chore(embedding-model): add v2 interface
+- 443d8ec: feat(embedding-model-v2): add response body field
+- 5c9eec4: chore(providers/anthropic): switch to providerOptions
+- 66962ed: fix(packages): export node10 compatible types
+- d9209ca: fix (image-model): `specificationVersion: v1` -> `v2`
+- 9301f86: refactor (image-model): rename `ImageModelV1` to `ImageModelV2`
+- 7378473: chore(providers/google): switch to providerOptions
+- 779d916: feat: add provider option schemas for vertex imagegen and google genai
+- 91715e5: fix (provider/google-vertex): fix anthropic support for image urls in messages
+- d1a034f: feature: using Zod 4 for internal stuff
+- fd65bc6: chore(embedding-model-v2): rename rawResponse to response
+- 205077b: fix: improve Zod compatibility
+- bb13f18: Add reasoning token output support for gemini models via Vertex AI Provider
+- Updated dependencies [a571d6e]
+- Updated dependencies [742b7be]
+- Updated dependencies [e7fcc86]
+- Updated dependencies [78e7fa9]
+- Updated dependencies [7cddb72]
+- Updated dependencies [ccce59b]
+- Updated dependencies [e2b9e4b]
+- Updated dependencies [95857aa]
+- Updated dependencies [45c1ea2]
+- Updated dependencies [6f6bb89]
+- Updated dependencies [ad66c0e]
+- Updated dependencies [f916255]
+- Updated dependencies [060370c]
+- Updated dependencies [dc714f3]
+- Updated dependencies [b5da06a]
+- Updated dependencies [8f2854f]
+- Updated dependencies [d1a1aa1]
+- Updated dependencies [63f9e9b]
+- Updated dependencies [5d142ab]
+- Updated dependencies [d5f588f]
+- Updated dependencies [e025824]
+- Updated dependencies [19a4336]
+- Updated dependencies [0571b98]
+- Updated dependencies [5d959e7]
+- Updated dependencies [8af9e03]
+- Updated dependencies [b6b43c7]
+- Updated dependencies [4fef487]
+- Updated dependencies [48d257a]
+- Updated dependencies [0c0c0b3]
+- Updated dependencies [0d2c085]
+- Updated dependencies [1a635b5]
+- Updated dependencies [40acf9b]
+- Updated dependencies [888b750]
+- Updated dependencies [9222aeb]
+- Updated dependencies [3259565]
+- Updated dependencies [8dfcb11]
+- Updated dependencies [9f73965]
+- Updated dependencies [e2aceaf]
+- Updated dependencies [411e483]
+- Updated dependencies [8ba77a7]
+- Updated dependencies [fdff8a4]
+- Updated dependencies [eb173f1]
+- Updated dependencies [4f26d59]
+- Updated dependencies [25f3454]
+- Updated dependencies [6a16dcf]
+- Updated dependencies [a85c85f]
+- Updated dependencies [7b3ae3f]
+- Updated dependencies [a166433]
+- Updated dependencies [26735b5]
+- Updated dependencies [5cf30ea]
+- Updated dependencies [443d8ec]
+- Updated dependencies [5c9eec4]
+- Updated dependencies [a8c8bd5]
+- Updated dependencies [abf9a79]
+- Updated dependencies [14c9410]
+- Updated dependencies [e86be6f]
+- Updated dependencies [9bf7291]
+- Updated dependencies [c68931f]
+- Updated dependencies [2e13791]
+- Updated dependencies [9f95b35]
+- Updated dependencies [66962ed]
+- Updated dependencies [0d06df6]
+- Updated dependencies [472524a]
+- Updated dependencies [dd3ff01]
+- Updated dependencies [a313780]
+- Updated dependencies [d9c98f4]
+- Updated dependencies [05d2819]
+- Updated dependencies [9301f86]
+- Updated dependencies [fd98925]
+- Updated dependencies [0a87932]
+- Updated dependencies [c4a2fec]
+- Updated dependencies [957b739]
+- Updated dependencies [79457bd]
+- Updated dependencies [a3f768e]
+- Updated dependencies [cb787ac]
+- Updated dependencies [7378473]
+- Updated dependencies [7435eb5]
+- Updated dependencies [f07a6d4]
+- Updated dependencies [8aa9e20]
+- Updated dependencies [4617fab]
+- Updated dependencies [075711d]
+- Updated dependencies [75f03b1]
+- Updated dependencies [779d916]
+- Updated dependencies [ac34802]
+- Updated dependencies [0054544]
+- Updated dependencies [269683f]
+- Updated dependencies [581a9be]
+- Updated dependencies [cb68df0]
+- Updated dependencies [ad80501]
+- Updated dependencies [68ecf2f]
+- Updated dependencies [9e9c809]
+- Updated dependencies [32831c6]
+- Updated dependencies [6dc848c]
+- Updated dependencies [6b98118]
+- Updated dependencies [d0f9495]
+- Updated dependencies [63d791d]
+- Updated dependencies [87b828f]
+- Updated dependencies [2e06f14]
+- Updated dependencies [3f2f00c]
+- Updated dependencies [d601ed9]
+- Updated dependencies [bfdca8d]
+- Updated dependencies [0ff02bb]
+- Updated dependencies [b9ddcdd]
+- Updated dependencies [91715e5]
+- Updated dependencies [7979f7f]
+- Updated dependencies [ca8aac6]
+- Updated dependencies [39a4fab]
+- Updated dependencies [44f4aba]
+- Updated dependencies [61ab528]
+- Updated dependencies [84577c8]
+- Updated dependencies [9bd5ab5]
+- Updated dependencies [57edfcb]
+- Updated dependencies [faf8446]
+- Updated dependencies [7ea4132]
+- Updated dependencies [8e6b69d]
+- Updated dependencies [42fcd32]
+- Updated dependencies [d1a034f]
+- Updated dependencies [5c56081]
+- Updated dependencies [fd65bc6]
+- Updated dependencies [023ba40]
+- Updated dependencies [ea7a7c9]
+- Updated dependencies [26535e0]
+- Updated dependencies [e030615]
+- Updated dependencies [6392f60]
+- Updated dependencies [878bf45]
+- Updated dependencies [5e57fae]
+- Updated dependencies [393138b]
+- Updated dependencies [c57e248]
+- Updated dependencies [0f05690]
+- Updated dependencies [7badba2]
+- Updated dependencies [88a8ee5]
+- Updated dependencies [41fa418]
+- Updated dependencies [205077b]
+- Updated dependencies [71f938d]
+- Updated dependencies [3795467]
+- Updated dependencies [28a5ed5]
+- Updated dependencies [7182d14]
+- Updated dependencies [ee5a9c0]
+- Updated dependencies [f418dd7]
+- Updated dependencies [c1e6647]
+- Updated dependencies [1766ede]
+- Updated dependencies [362b048]
+- Updated dependencies [399e056]
+- Updated dependencies [0b678b2]
+- Updated dependencies [811dff3]
+- Updated dependencies [f10304b]
+- Updated dependencies [dd5fd43]
+- Updated dependencies [a753b3a]
+- Updated dependencies [33f4a6a]
+- Updated dependencies [383cbfa]
+- Updated dependencies [27deb4d]
+- Updated dependencies [c4df419]
+  - @ai-sdk/provider-utils@3.0.0
+  - @ai-sdk/provider@2.0.0
+  - @ai-sdk/google@2.0.0
+  - @ai-sdk/anthropic@2.0.0
+
+## 3.0.0-beta.21
+
+### Patch Changes
+
+- Updated dependencies [88a8ee5]
+  - @ai-sdk/provider-utils@3.0.0-beta.10
+  - @ai-sdk/anthropic@2.0.0-beta.13
+  - @ai-sdk/google@2.0.0-beta.19
+
+## 3.0.0-beta.20
+
+### Patch Changes
+
+- Updated dependencies [78e7fa9]
+- Updated dependencies [0f05690]
+- Updated dependencies [f418dd7]
+- Updated dependencies [27deb4d]
+  - @ai-sdk/google@2.0.0-beta.18
+  - @ai-sdk/anthropic@2.0.0-beta.12
+  - @ai-sdk/provider@2.0.0-beta.2
+  - @ai-sdk/provider-utils@3.0.0-beta.9
+
+## 3.0.0-beta.19
+
+### Patch Changes
+
+- Updated dependencies [eb173f1]
+- Updated dependencies [dd5fd43]
+  - @ai-sdk/anthropic@2.0.0-beta.11
+  - @ai-sdk/google@2.0.0-beta.17
+  - @ai-sdk/provider-utils@3.0.0-beta.8
+
+## 3.0.0-beta.18
+
+### Patch Changes
+
+- Updated dependencies [e7fcc86]
+- Updated dependencies [269683f]
+  - @ai-sdk/provider-utils@3.0.0-beta.7
+  - @ai-sdk/anthropic@2.0.0-beta.10
+  - @ai-sdk/google@2.0.0-beta.16
+
 ## 3.0.0-beta.17
 
 ### Patch Changes